requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cloudinary==1.43.0
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
